Topology connectivity enforced
Used the connected_components_subgraph command in networkx to determine
if multiple components existed between the collection of bonds to be put
into a topology.

Perhaps I am being too strict with what a "Topology" is. In actuality
topology is the study of the topological properties of a graph and a
graph need not be connected. Perhaps it would be better to stick with
the other libraries choice of "Topology" meaning the whole system and I
type-check that a MoleculeTopology is indeed fully connected.