System constructor working, enforces non-overlap
Had to add the Molecule.overlaps method and the overlaps function in the
system module.

Could later implement a Molecule.overlap that actually returns the
overlapping atoms as pairs.

Could also implement a more sophisticated function to check if your
system's atoms are in a state which won't be good for dynamics or
whatever.

Such as atoms being too close, but not necessarily overlapping.