First commit with new Selections module.
Lots of improvements to the base datastructures classes.

SelectionMember supports having a 'registry' attribute which if used
by the selection class (currently IndexedSelection does this) will
associate a selection to the SelectionMember through the global
selection_registry. This allows a SelectionMember to keep track of which
selections it is a part of and to access those selections.