Added SelectionDict and SelectionList classes
These just take UserDict and UserList and cross inherit them with
SelectionMember so that they can be selected from as well.

Only enforce that input datastructures be subclasses of
collections.Mapping and collections.Sequence respectively.

Maybe should enforce that things in these structures also be
SelectionMembers?

Will wait for case that this occurs.