RDKitMoleculeType.to_molecule working
Using the atoms_constructor for Molecule an RDKitMoleculeType can create
a 3D molecule from one of it's conformers.

It's a little messy perhaps as right now I have two constructors, atoms
constructor and type_constructor. The idea is to have the atoms
constructor as the nuclear option where you create each atom
individually. The other is to simply to have a type constructor where
you hand it a type and coordinates for it's atoms and it makes an
molecule for you. This is a bit harder because you have to map the atoms
and coordinates and having the interface for both in the constructor
will just take some finangling. For now it is okay because it works for
the use-case I need it for.