AtomType, AtomTypeLibrary, rdkit atom_data method
Made a function to return a dictionary for relevant rdkit atom data.

The AtomType class consumes dictionaries like this and adds them to it's
attributes through it's __dict__ attribute. Not sure if this is a kosher
way of doing things but it seems to work.

The AtomTypeLibrary was made to store the different AtomTypes without
duplicating them. Needed comparison operators implemented for AtomType
which is done. Method to add an AtomType and a method to check whether
the AtomType's attributes match any in the library, in case they have
different names.

When coming up with an example I had the problem of having atoms that
are different (e.g. bond degree) and have the same pdb_name (which I was
using for the atom_name in the AtomTypeLibrary dict. So I wrote a
functino that checks the library and appends numbers to each atom name
if there is an atom that is different of the same element symbol. Should
put that somewhere for safe-keeping as it will probably be
useful. Although it is important to realize it is external to the
library itself, unless put in a specially named function and that I
should resist temptation to bake it into the library class.