Bond class and bonds attribute to Atom
Created the Bond class.

Atom is now aware of the bonds it is a part of (through both it's
registry and the _in_bond bool attribute) and can access them through
the bonds attribute.

Found a very strange bug where isinstance(Molecule(), Bond) evaluates to
True when implementing the bonds attribute getter. Haven't tested
elsewhere but the Molecule returned by get_selections needs to be
filtered out by testing to make sure the selection is not a Molecule.