Atom.adjacent_atoms method attribute
Added a method to retrieve the adjacent atoms.

I wanted to have it so yo could define how many bonds away you wanted
atoms but this was too difficult to implement and I only need the ones a
single bond away for now. Later I can add that feature in if I need to.