Interaction Profiling almsot complete
Got arbitrary InteractionTypes working for a single molecule profiling.

Works for HydrogenBondType and returns HydrogenBondInx objects.

Had to take out the part where it checks that the atoms are in the
system as it was giving me a ton of trouble.

Introduced an InteractionError to help with debugging.

__main__ scripts updated to include testing of H-bonds within a single
molecule only.