System introspection implemented
Implemented methods similar to Atom's method for finding out if it is in
a Molecule.

Had to update SelectionList to correctly assign it's members registries.

The need for a SystemMember class became apparent. This would be a
mix-in that would simply provide the _in_system attribute and the system
method for finding if a Member is in a system by looking recursively in
it's selections.