clear_registry, inx templates, parameters, example
clear_registry: I needed to have control over deleting things out of the
SELECTION_REGISTRY as it was causing a huge memory leak over multiple
iterations. So I made the clear_registry function. Probably a better
method is needed to be implemented in the future.

inx templates: a template class for defining the InxType and the
actual Inx object for all the different kinds of interactions.

parameters: copy-pasted and formatted the parameters for interactions
from PLIP and put them into the interactions module.

example: I updated a number of the scripts in the modules if __name__ ==
"__main__" sections so that they worked properly.

* TODO:

- make proper Type classes for System and SystemAssociation
- documentation
- logging
- selection references out of global variable