added modules: data, interfaces.rdkit
Added the data module in the test module. When imported the __init__
file takes care of loading the files as strings. Probably should make
this JIT so it doesn't take forever when we add more.

moved the RDKitWrapper to a separate module interfaces.rdkit.

I added the interfaces module in order to allow for the addition of
other wrappers and writers in the future.

Started writing tests for RDKitWrapper.