RDKitMoleculeWrapper.find_features position, tests
Changed the behavior of this to not return an rdkit object that has a
bunch of methods for calculations and rather just the x,y,z coordinates.

Now it just gives you a float (x,y,z) tuple.

Most of the tests (maybe they are trivial, oh well) for RDKitWrapper are written.