Type constructors, SelectionMember flags
Finished the type constructors for the three main molecule.py classes.

Also added the flags attribute to SelectionMembers so that they can be
flagged as 'molecule' or 'system' to indicate that they are apart of a
selection that is those. Of course the interpretation of the string
flags is up to whatever method. Will add isin_molecule etc methods later
that looks for this.

Cleared out some of the unnecessary SelectionType stuff in selection.py
as well.

SelectionDict and SelectionList are renamed to SelectionsList and
SelectionsDict to reflect the fact that they are a selection but
propogate their selection into the containers of (possibly) selections
that they have.