interactions modularized, records and df functions
Moved interactions to a module itself, that way the file wasn't crowded
with new InteractionTypes, and they can be added by just adding a file.

Will have to play with __init__ to get them in the namespace nicely.

Got distracted and started adding functions for producing records for
each type.

These can be used to make dataframes which functions were also written.

Finished them for molecule.py and features.py