Writing pi-stacking, added example molecules
Am having trouble identifying aromatics so I added some examples of
loading a protein and ligand with and without Hydrogens. The protonated
trypsin gives errors in RDKit and thus cannot be sanitized, which is
needed to detect Aromatic features. An unprotonated version solves this
problem. However, the same is not true of benzamidine, which does not
have it's aromatic ring detected as a feature even when it is sanitized.