Examples, map pdb to mol
Mostly adding files to the mast.tests.data module.

Additionally I added a modified function from rdkit/Chem.py for
assigning bond orders from a template molecule to a target. This is
useful for loading in coordinates from a pdb small molecule and
detecting the features (aromaticity) from a proper molecule format like
a .mol file. I was getting errors in the SanitizeMol from the rdkit
version so I copied it and put it in the interfaces.rdkit module and
removed the SanitizeMol function.