Fixed registry flags, distance functions
SelectionsList and SelectionsDict were not passing on flags for
selections from higher levels. I fixed this so that the flags are passed
to their members. i.e. when you put a molecule in a system the atoms get
the 'system' flag. However, this does not make the system selection show
up in the registry of the atoms. Thus, to get the system you have to
excplicitly call for the system of the molecule this atom is a part
of. THat is unless the atom is not in a molecule and is a full-fledged
member of the system. THis may change in the future, but it works for
now.

Worked on the within_distance function for atoms and it is working.