Added atom_idxs to within_distance for molecule
THis allows you to select specific atoms from the molecule to compare
distances to. This is easier than just getting the atoms and using the
atoms_within_distance because that would give you atoms on the same
molecule. Next I would just want to be able to compare the self molecule
to any specific system member.