Fixed member feature pairings in profiling inxs
Also fixed the atoms attribute in the MoleculeAtomSelection so that it
was only the values of the dict of atoms! Was confusing before, and also
gives a better reason for the attribute instead of an alias for the data.

Had a problem where when I flipped the order of the members in find_hits
for the hydrogen bond part it gave them back without the information of
how they werer flipped.

When the profile_interactions received the feature pair tuples it
assumed the first was member_a and the second was member_b, however the
first was always the donor and the second the acceptor, which obviously
could be either member. So I returned the member order information as a
tuple and handled that appropriately in profile interactions.