Allow passing in of geom. params. for profiling
Previously it was hardcoded into the HydrogenBond classes which values
to use for checking for h-bond existence. So the infrastructure of
higher classes didn't allow for passing them in. This is changed and in
the HydrogenBondType class the defaults are set to the config file
values which were previously hardcoded.