__hash__ fix, inx class classification performance
__hash__ needs an integer return value so I used the hash of the name
not just the name.

The way I had the interaction classes applied to new interaction hits
was really, really slow because it unpacked all of them (like 100k in
this case), made the match (by comparing feature pairs) then passed that
into the constructor. Ya wya too much overhead, but I was trying to
avoid having to use the setter for the object. I changed it so that
find_hits only tries to find the appropriate interaction class after it
has the hit and then uses a while loop to find the matching interaction
class, so that iteration over all is avoided.

This increases performance to a normal level of dynamic interaction
class assignment.