Reworked find_hits to be general in interactions.py
Took most of the boilerplate stuff from the find_hits function in each
individual InteractionType subclass and converted it to general code in
the parent class. This needs to be tested still.

This dramatically reduces boilerplate for writing InteractionType
subclasses and will help with expanding the repertoire of them.

Also added tangling of scripts for setting up and saving a system as a
pickle.

Added a system pickle to the data.