Reworked feature keywords and classifiers
Earlier I had an implementation where the separate feature
classification terms for a feature were dumped into it's namespace and I
had to recover them later and figure out which one I wanted to
use. However I decided this was unsustainable and I want to have them
all in a common attribute. The drawback to this is name collisions,
however I think it won't be so hard to come up with a way to uniquefy
them and have InteractionTypes recognize them through the config
fil. That way the class source does not have to change!

I did this for the hydrogen_bond module and changed the rdkit interface
wrapper to put them into a controlled attribute defined in the config
file for features. I still could dump the attributes into the namespace
as 'metadata' but I forgot to and it is not that important I don't think
right now.

This resulted in having to change some of the logic in the interactions
submodule for matching interaction classes and I think it is for the
better now though.

The class attributes are much cleaner now for InteractionTypes and this
is reflected as well in the config file.