Fixing bugs put in from last commit
Profiling is working again, with all the improvements to clarity from
the config file stuff including classifiers etc.

This should make the code very portable. the test will be writing the
HalogenBond which should only require changing the class attributes and
making sure the config file is correct.

Ideally this could be done without writing any class HalogenBondType
code.

like:
>>> HalogenBondType = copy(HydrogenBondType)
>>> HalogenBondType.interaction_name = "HalogenBond"
etc...

IDK if that will work but this could be a powerful to make have a
template library for users to use and then they could customize their
own classifiers etc.