Cleaned up pistacking test functions, centroid rejection
Did more logical functionalization of the functions needed to calculate
things about pi stacking.

Implemented a proper vector rejection to get the projection onto the
plane of the other ring for either centroid.

Added some layers so that the functions are usable on raw coordinates
and the whole atoms/features is handled elsewhere.