Added record methods
For Interaction, InteractionType, Association, Associationtype, Feature, FeatureType,
System, and SystemType are now implemented.

Although thie info in the substantiated ones might not be enough.

Right now it just gives the name of their type, and for Interaction
the parameters from the geometry matching.

This can be used, (perhaps annoyingly) to get more info, assuming you
know which collection to get that info from.