gen_inx_space, unit assocs, profile_inxs, ex
Finished debugging generate_interaction_space and the
unit_associations methods.

I also removed a lot of cruft from profile_interactions that was left
from before I started using the association terms approach. Before I
had it do the forward and backwards part of assymetric
interactions. This made an extra nested loop and some checks and stuff
at the beginning. That is all gone now and the method is significantly
simpler.

The onus is now on the invoker of the method to provide exactly the
members they wish to do the profiling on. I.e. (0,1) will not give
you (1,0) under any circumstances in profile_interactions and it is
not optional.

I also made some changes and divied up some stuff in the TPPU sEH
example and made some coordinate npy files for use around the block.