Working profile_interactions module
Implemented a model similar to scikit-learn objects where you make the
model object first and then you fit data data to it later.

Meaning I start with making a SystemType giving it an InteractionSpace
then substantiating it to a System then make a SystemProfile object
which takes only the system as an argument. From here we profile the
associations within the SystemProfile individually. At the end we have
an object holding all of the results from the profilings inside
it. Underneath it holds each associations profiling results as a
separate AssociationProfile object in an OrderedDict, so that they are
indexable and accessible with the association term (0,1) etc. There
are methods to get all the hits out at once, because they are unique
within the InteractionSpace anyhow.

The trade-off with the current implementation is that you are stuck
with the InteractionSpace you generated after making a SystemType. I
think this could be improved but the complicatedness was keeping me
from focusing on the short term goal. If the need arises we could
consider changing that. It shouldn't be too difficult to add
complexity ones the core workflow is determined and tested.

ANother concern is that this is only implemented for unit associations
for the time being. Again this shouldn't be too difficult to add once
I am happy with the current workflow.