Reworking Profiling.
Big change structurally for how to do profiling.

I have made it so that interaction spaces are separate from
association_types and system types. But you use these to generate
interaction spaces.

There are a few reasons for this.

The SystemType would become very bloated with all of those things, and
it is not clear that there should only be one Interaction space per
SystemType. So while there is a possible limit of the number of
interaction classes for a system, given a finite number of interaction
types, you could have different sets of space you want to treat
separately.

So the workflow for creating an interaction space is to add subspaces
one at a time, where a subspace maps one-to-one to AssociationTypes.

When you add a subspace it takes the AssociationType and an
InteractionType and generates interaction classes from that. Those
interaction classes are added to the InteractionSpace and the indices
of them are tracked. This allows for indexing of interaction classes
in the space (for hit_idxs) and for mappings of subspace keys to their
interactions, in the form of (AssociationType, InteractionType).

The second benefit of having it this way is that we can now have a
model where we profile an interaction space (since an interaction
space has exactly one SystemType) over substantiations of it's
SystemType.

So I introduced the InxSpaceProfiler which takes only an
InteractionSpace to construct it. The InxSpaceProfiler then generates
InxSpaceProfiles from the profile method which takes in Systems
substantiated from the SystemType.

This gives a very clean way of creating a "model" (the
SystemType + InteractionSpace) for which you feed in data and get
results out of.

So far the Profile objects are not finished but have basic
functionality. In future commits they will have added features for
doing analysis more easily and being able to dispose of the
Interaction object, which are kind of sticky objects and tend to take
up a lot of memory (I think).