Circumvent problem with multiple hit_idx assignment
It is probably too complicated to have the hit records by inx_type so
I just made it so that you can output records and dfs from the
interactions using just there Interaction class record making
function.

You can assign hit indices after the fact, or this can be added to
this function.

Will test adding hit_idx manually after to test but then add it in to
these method.