Made profiling in order of interaction_space
I may have introduced a bug where I was looping over the interactions
batching by the subspace they were in, which was a dictionary.

Because order is not guaranteed this may have caused problems.

However, I don't think this caused the behavior I am seeing now.

It is a good thing to catch this though.

Now SystemProfile constructor simply uses the order of the interaction
classes in the order of the inx_space.