removed the member attribute of SelectionMember
This results in a reference of these objects to itself.

Perhaps this could have been designed better but the data of the
member is not actually handled in the ABC SelectionMember, but in the
subclass where that data is specific to the real class.

So for this I had to change the CoordArray to hold it's own data
attribute much like the IndexedSelection and SelectionDict and
SelectionList.

Everything seems to work. I grepped the project and nowhere else was
that attribute used.