fixed interaction class creation for associations
This was pretty involved.

I discovered that the AssociationTypes were not really storing data
for the selections.

ANd further there were problems with the making of the selections.

So I updated the MoleculeAtomSelection so it was split into
MoleculeAtomSelection and MoleculeTypeAtomSelection, for obvious
reason, even though the code is kind of copied.

I also added the option of whether to include partial features and
bonds or not in this selection. Which is passed through the
AssociationType constructor through a kwarg that gets mapped to the
constructor of the selections.

Then in the InteractionType.interaction_classes it just iterates
through the elements of the association (set by the superclass
constructor, accessible through __getitem__).

That means that now the AssociationTYpe is a subclass of the
SelectionsList.

And also the MoleculeSelection needed that the Molecules, Atoms, etc
were also from SelectionMembers, so I did that.

Which meant that the system pickle had to be remade.

I should test to make sure that the No selection is working for
selecting members