mucked with the records dataframe methods to make them easier
There was a strong need to incorporate the information about the
identity of the atoms in the interactions.

So I went and did it and relied on the PDB stuff and also dumped all
the interaction_class fields in the interaction records.

But the records were gettinga bit stifling so I just with a dictionary
approach.

THis all should be redone in a clean way in which the (truly optional)
PDB stuff is done in a special method that knows what to expect for
that.

Later however we will have our own residue protein system to contend
with so we can just wait until then and leave this very useful
behavior where it is for now.